refactor: move annotation app platform settings to separate file
